Updated for prime tables starting with 1 and 2
Replaced unsigned by size_t for table sizes

+++ prime.h
@@ -16,9 +16,9 @@ Compiler        : Mingw-w64 g++ 11.2.0
 #define LABO06_C_BA_MCRM_PRIME_H
 
 /// Function to remove all non primes numbers from a table
-/// @param table Table to wipe from 1 to Size
+/// @param table Table to wipe from 1 or 2 to Size
 /// @param tableSize Size of the table
 /// @return Size of the table after wipe
-unsigned primeTable(unsigned table[], unsigned tableSize);
+size_t primeTable(unsigned table[], size_t tableSize);
 
 #endif //LABO06_C_BA_MCRM_PRIME_H
